Scene*: Move shader creation/destruction to setup()/teardown().
Up to now, when the shaders used for a scene were option-independent,
the shader creation/destruction code was placed in load()/unload(). For
the upcoming shader precision options to work, however, this can no longer
be done: all shaders are dependent on the precision options and should
therefore be placed in setup()/teardown().