Android: Use a custom EGLConfigChooser that gracefully handles eglChooseConfig failures.
When proper GLES2.0 drivers are missing from an Android device,
eglChooseConfig() will fail to find a GLES2.0 capable config, as expected.
Unfortunately, the default EGLConfigChooser implementation doesn't give us a
chance to handle this error; it throws an exception that we cannot catch
(because it comes from an internal thread) and crashes the application.

In order to handle the error gracefully, this commit introduces a custom
EGLConfigChooser implementation that in case of error displays an informative
alert and then quits the application.