Add --data-path command line option
glmark2 needs access to models, shaders and textures.  These are stored
in the glmark data path.

Currently glmark2 has a --data-path build-time option to specify where
these files are installed at build time.  This option is passed in at
compile time via the GLMARK_DATA_PATH macro.

In some situations, however, it might become necessary to specify the
location of the data files at *runtime*.  This patch adds that option.
If not specified, the compile-time GLMARK_DATA_PATH value is used.

In particular, this feature is used when using glmark2 as a Chrome OS
autotest dependency.  glmark2 is built and installed in the default
locations in a development host chroot (ie ${SYSROOT}/usr/bin/glmark2 and
${SYSROOT}/usr/share/glmark2).  The autotest deps system then rsyncs the
executable and data into a special sandbox on the DUT.  Thus, the data
files will not actually get installed at /usr/share/glmark2, and instead
we use this option to specify their location within the autotest sandbox.

Signed-off-by: Dominik Behr <dbehr@chromium.org>