NativeStateX11: Don't terminate --fullscreen if EWMH Atom not found
If the Atom "_NET_WM_STATE_FULLSCREEN" is not defined in the X server, it
could mean that the active Window Manager is not EWMH compliant, or that
there is no window manager at all.  In this case, the XChangeProperty
will fail and the program will terminate with the message:

X Error of failed request:  BadAtom (invalid Atom parameter)
  Major opcode of failed request:  18 (X_ChangeProperty)
  Atom id in failed request:  0x0
  Serial number of failed request:  28
  Current serial number in output stream:  30

Instead, if the Atom doesn't exist, don't call XChangeProperty, just log
a message and keep going.  The window will still be fullscreen, it just
may still have decorations.

Fixes: https://launchpad.net/bugs/1270871