wscript: check C libraries using the C compiler
On recent g++ versions, the configure stage will fail when checking for
the sqrt function with the following compiler error:

../test.cpp:11:16: error: overloaded function with no contextual type information

This is due to the fact that 'sqrt' also exists in the std namespace. As
the functions tested are C functions, use the C compiler to check for
their presence in order to avoid these issues.