Bypassing the module checking system again, with udev this time.
Using udev to determine the main graphic card DRM device node, and use
it instead of querying every potential past, present, future and
home-made DRM modules names, as it makes no sense.

Additional corner cases will be dealt with the addition of an option
to force the DRM device node to use, instead of querying it, in a future
patch (aka, it's on the TODO list).

Signed-off-by: Myy <myy@miouyouyou.fr>