Add Win32 EGL flavor.
Updates the build files to add a 'win32-glesv2' flavor target. We
can use this flavor to load ANGLE on Windows where there are no
native GLES/EGL drivers.

Falls back to eglGetDisplay when there is platform enum. This is a
generalization of the DispmanX config. Some EGL platforms such as
ANGLE don't support any of the currently listed enums. In lieu of
using the special enum for ANGLE we can fall back to eglGetDisplay.

Necessary for Windows support with ANGLE (#9).