Allow flavours to generate their own source files
Wayland protocol extensions are handled by a central distribution of XML
files which describe the new protocol extension. Each user of the
extension then runs the wayland-scanner tool over the XML to generate C
source and headers, or another scanner tool for another language.

This two-stage build means we need to add support to waf for generated
sources.