glad: Update for GL 2.1 + GL_EXT_framebuffer_object
Update glad to be based on GL 2.1 with GL_EXT_framebuffer_object instead
of GL 3.3. This allows us to run more scenes on systems that have older
GL versions and ensures we don't inadvertently use any newer features.

This requires using the EXT versions of the GL framebuffer functions in
OpenGL builds.

Fixes #124