fix benchmark parameter edition
when a parameter of a benchmark is edited and this parameter prompts a
dialog with a EditText, there is no way to validate the new value as the
"Done" button in the virtual keyboard doesn't appear and the "Enter"
button creates a newline in the EditText widget.

this was fixed by forcing the EditText to be single-line.