NativeStateGBM: Add GBM offscreen backend
glmark2 supports offscreen runs and even GBM already in the case of the
DRM backend. However, in all cases even with the --off-screen flag, it
goes through display or a window system initialization which requires
connected hardware or additional environment setup to satisfy that
initialization (connected monitor or Xorg or a Wayland compositor
running).

This commit adds a simple GBM backend which is purely offscreen and
requires no display or window system setup. The advantage is being able
to run on systems with no display-related requirements, and providing
GPU benchmark results completely independent of displays. It is of
course not possible to visually see the results, as is already the case
with other offscreen runs.

The benchmark results with are equivalent, for the same surface size, to
a run with other backends that work in a similar fashion, so compatible
with current glmark2 scores. It is also possible to run with the
--off-screen flag for the same alternate offscreen framebuffer path as
other backends.

This backend can be helpful in, for instance: development systems,
automated test runs, CI environments, or runs on headless remote
systems.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>