Scene: Introduce non-virtual methods to prepare and finish a scene for benchmark
Introduce the 'prepare' and 'finish' non-virtual methods, and use these
for benchmark scene preparation and finish. Also make Scene load/setup
and unload/teardown methods protected to disallow the main benchmarking
code from using these directly.

The purpose of this change is so that in upcoming commits we are able
to:

1. Easily add setup/teardown behavior that applies to all scenes.
2. Relieve the per-scene setup/teardown code from having to manually call
   into the base class.