GLStateEGL,GLStateGLX: Fix the failure to find suitable FB config
It does not make sense that ``select_best_config()`` returns NULL
from matched FB configs even if there is only one config which gets
a negative score in the list. For example, on my WSL2 Ubuntu-20.04,
that situation mentioned occurs.

    Debug: Found 1 matching FB configs.
    Debug: matched FB config score -785.
    Debug: GLX chosen config ID: 0x23 Native Visual ID: 0x23
    Debug:   Buffer: 24 bits
    Debug:      Red: 8 bits
    Debug:    Green: 8 bits
    Debug:     Blue: 8 bits
    Debug:    Alpha: 8 bits
    Debug:    Depth: 0 bits
    Debug:  Stencil: 0 bits
    Debug:  Samples: 0
    Debug: Creating XWindow W: 800 H: 600 VisualID: 0x23

And with this patch applied, glmark2 will run like a charm for me.

Signed-off-by: Luc Ma <luc@sietium.com>