x11: Correctly set fullscreen size using Xrandr
When running `glmark2 --fullscreen` on a multi-screen setup, the glmark2
surface size would be the total screen size, causing fullscreen display issues.
For example, when two monitors are connected in extended mode with resolutions
of 1920x1080 each, the surface size would be 3840x1080 fullscreen.
This would result in the glmark2 window only partially displaying on the primary
monitor, and not showing on the secondary monitor at all.
This modification ensures that the fullscreen size is set to the primary monitor's
resolution instead.

Signed-off-by: likai <likai@kylinos.cn>